add a check on T_CreateFunctionStmt
Shir Tamari (Wiz) identified that functions with the language type 'internal' could be created to call the built-in C language functions
directly. This would allow bypassing the disallow-list of builtin functions.

The library/agent now prevents creating functions of language type 'internal' when in an elevated context. This language type is usually disallowed for untrusted/not-superuser roles, thus the change should be safe. We can't block LANGUAGE c as this is legitimately called by some extensions (pg_repack etc).