protect columns in pg_proc, block additional pg_ table modification
We can't just block pg_proc modification as there are official extensions that modify pg_proc to some degree during installation or upgrade.
This approach checks which columns are being modified and blocks if a "sensitive" column is altered.

Also adds the pg_authid and pg_auth_membership to the tables that aren't allowed to be modified.