allow defining a list of allowed superusers
Introduces the `aiven.pg_security_agent_reserved_roles` config option.
A comma separated list of roles that are allowed to be assigned the superuser
property can be defined. Trying to grant superuser to any role not in the list,
regardless of the 'context' (elevated or extension install) will be denied.